Move StrEnum into utils
